+++ crates/module-system/module-implementations/sov-prover-incentives/src/capabilities.rs
@@ -3,10 +3,9 @@ use std::cmp::max;
 use sov_bank::{config_gas_token_id, Coins, IntoPayable};
 use sov_modules_api::registration_lib::StakeRegistration;
 use sov_modules_api::{
-    AggregatedProofPublicData, Gas, InvalidProofError, SerializedAggregatedProof, Spec,
-    StateReader, TxState, Zkvm,
+    AggregatedProofPublicData, Gas, InvalidProofError, SerializedAggregatedProof, Spec, TxState,
+    VersionReader, Zkvm,
 };
-use sov_state::User;
 use thiserror::Error;
 
 use crate::event::SlashingReason;
@@ -260,7 +259,7 @@ impl<S: Spec> ProverIncentives<S> {
     }
 
     /// Check that the initial and final state values of the proof output are valid against the chain state module
-    fn check_proof_outputs<ST: StateReader<User>>(
+    fn check_proof_outputs<ST: VersionReader>(
         &self,
         public_outputs: &AggregatedProofPublicData,
         state: &mut ST,